requirements.txt
Flask==3.0.3
pandas==2.2.2
xlsxwriter==3.2.0
reportlab==4.2.2
openpyxl==3.1.5
arabic-reshaper==3.0.0
python-bidi==0.4.2
gunicorn==21.2.0
